Rename `acvm_interop` -> `backend`
